🚀 Day 22: Reactor Reboot
Part 1 was easy.

For part 2, it is obvious no brute force (as in part 1) will work in finite time. Standard structure won't fit the sample puzzle (int arrays), regardless the actual puzzle. A more complex structure would be possible in theory, but it an obvious fact that AoC won't require that...

The initial approach was working with distinctions of cubes (cuboid), which means for every new cube (of an instruction) the existing cubes will be cut. However, at the end the calculation won't work out (missed all correct distinctions). Also this results into some complex algorithm and explodes the required cubes (because they will be cut/split several times again and again) :/

The approach with intersections and a weight (https://github.com/tjol/advent-of-code-2021/blob/main/22/puzzle2.js) is far better: it also creates additional volumes, but the existing ones will be weighted. The actual trick is the "always-remove" phase: Even a "state=on" cube will create a negative cube opposite.

The actual sum is the result of all volumes.. at least code I had already due my previous approach.