🧮 Day 24: Arithmetic Logic Unit
Initial approach was using a standard algorithm. However, this will not work out for 9^14 dimensions.

After that, looking for optimizations in the instructions were.. boring. :)

So, at least the 9x9x9.. operations must be done only at once (my initial approach missed that obvious one). And after looking for some inspirations in the solutions thread, I found https://github.com/Mahrgell/AoC2021/blob/main/aoc21-24/src/main.rs which was described at reddit that it fold up the options (for part1 the max, for part2 the min). This means not following (sub)branches which are worse that already known ones. Sounds good enough for me.

So, this solution does not change and modify the input/instructions. Said that, it takes some minutes for the result.